+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Versioning].
 
 ## [Unreleased]
 
+- Change the deprecation warning to a deprecation error. ([#949])
 - Update transitive dependency `undici`. ([#948])
 
 ## [4.0.8] - 2024-02-17
@@ -636,5 +637,6 @@ Versioning].
 [#936]: https://github.com/ericcornelissen/svgo-action/pull/936
 [#937]: https://github.com/ericcornelissen/svgo-action/pull/937
 [#948]: https://github.com/ericcornelissen/svgo-action/pull/948
+[#949]: https://github.com/ericcornelissen/svgo-action/pull/949
 [64d0e89]: https://github.com/ericcornelissen/svgo-action/commit/64d0e8958d462695b3939588707815182ecc3690
 [8d8f516]: https://github.com/ericcornelissen/svgo-action/commit/8d8f516583b4340f692e2ea80e1855e5a1211bd3